Split history store into comment and history store
