+++ app/src/stores/CommentStore.js
@@ -1,18 +1,90 @@
 import { makeAutoObservable } from 'mobx';
+import axios from 'axios';
+import { format } from 'date-fns';
 
 export class CommentStore {
     editMode = false;
     editCommentIndex = null;
     editedCommentContent = null;
     isCommentListVisible = true;
+    commentTrigger = true;
 
     constructor(store) {
         this.store = store;
         makeAutoObservable(this);
     }
 
-    setEditMode = value => (this.editMode = value);
-    setEditCommentIndex = value => (this.editCommentIndex = value);
-    setEditedCommentContent = value => (this.editedCommentContent = value);
-    setIsCommentListVisible = value => (this.isCommentListVisible = value);
+    setEditMode = val => (this.editMode = val);
+    setEditCommentIndex = val => (this.editCommentIndex = val);
+    setEditedCommentContent = val => (this.editedCommentContent = val);
+    setIsCommentListVisible = val => (this.isCommentListVisible = val);
+    setCommentTrigger = val => (this.commentTrigger = val);
+
+    addComment = async comment => {
+        const comment_time = format(new Date(), 'H:mm do MMM yyyy OOOO');
+
+        const params = {
+            study_uuid: this.store.core.studyUuid,
+            user_uuid: this.store.core.userUuid,
+            history_item_index: this.store.core.studyHistoryItemIndex,
+            comment: comment,
+            comment_time: comment_time
+        };
+
+        this.store.core.studyHistory[
+            this.store.core.studyHistoryItemIndex
+        ].comments.push({ comment: comment, time: comment_time });
+
+        await axios.post('history/comment/', params);
+        this.store.history.generateHistoryNodes();
+
+        if (!this.store.core.studyIsSaved) {
+            this.store.core.saveStudy();
+        }
+    };
+
+    deleteCommnet = async index => {
+        const params = {
+            study_uuid: this.store.core.studyUuid,
+            user_uuid: this.store.core.userUuid,
+            history_item_index: this.store.core.studyHistoryItemIndex,
+            comment_index: index
+        };
+
+        this.store.core.studyHistory[
+            this.store.core.studyHistoryItemIndex
+        ].comments.splice(index, 1);
+
+        await axios.post('history/comment/delete', params);
+        this.store.history.generateHistoryNodes();
+    };
+
+    editComment = async (comment, index) => {
+        const comment_time = format(new Date(), 'H:mm do MMM yyyy OOOO');
+
+        const params = {
+            study_uuid: this.store.core.studyUuid,
+            user_uuid: this.store.core.userUuid,
+            history_item_index: this.store.core.studyHistoryItemIndex,
+            comment_index: index,
+            comment: comment,
+            comment_time: comment_time
+        };
+
+        this.store.core.studyHistory[
+            this.store.core.studyHistoryItemIndex
+        ].comments[index].comment = comment;
+
+        this.store.core.studyHistory[
+            this.store.core.studyHistoryItemIndex
+        ].comments[index].time = comment_time;
+
+        this.store.core.studyHistory[
+            this.store.core.studyHistoryItemIndex
+        ].comments[index]['edited'] = true;
+
+        await axios.post('history/comment/edit', params);
+
+        this.store.history.generateHistoryNodes();
+    };
 }
