+++ app/src/stores/HistoryStore.js
@@ -1,6 +1,5 @@
 import { makeAutoObservable } from 'mobx';
 import { MarkerType } from 'react-flow-renderer';
-import { format } from 'date-fns';
 import dagre from 'dagre';
 import axios from 'axios';
 
@@ -9,7 +8,6 @@ export class HistoryStore {
     edges = [];
     nodeHeight = 100;
     nodeWidth = 200;
-    commentTrigger = true;
 
     constructor(store) {
         this.store = store;
@@ -36,8 +34,6 @@ export class HistoryStore {
 
         return title;
     };
-
-    setCommentTrigger = val => (this.commentTrigger = val);
 
     generateHistoryNodes = () => {
         this.nodes = [];
@@ -154,74 +150,6 @@ export class HistoryStore {
 
         this.nodes = [...layoutedNodes];
         this.edges = [...layoutedEdges];
-    };
-
-    addComment = async comment => {
-        const comment_time = format(new Date(), 'H:mm do MMM yyyy OOOO');
-
-        const params = {
-            study_uuid: this.store.core.studyUuid,
-            user_uuid: this.store.core.userUuid,
-            history_item_index: this.store.core.studyHistoryItemIndex,
-            comment: comment,
-            comment_time: comment_time
-        };
-
-        this.store.core.studyHistory[
-            this.store.core.studyHistoryItemIndex
-        ].comments.push({ comment: comment, time: comment_time });
-
-        await axios.post('history/comment/', params);
-        this.generateHistoryNodes();
-
-        if (!this.store.core.studyIsSaved) {
-            this.store.core.saveStudy();
-        }
-    };
-
-    deleteCommnet = async index => {
-        const params = {
-            study_uuid: this.store.core.studyUuid,
-            user_uuid: this.store.core.userUuid,
-            history_item_index: this.store.core.studyHistoryItemIndex,
-            comment_index: index
-        };
-
-        this.store.core.studyHistory[
-            this.store.core.studyHistoryItemIndex
-        ].comments.splice(index, 1);
-
-        await axios.post('history/comment/delete', params);
-        this.generateHistoryNodes();
-    };
-
-    editComment = async (comment, index) => {
-        const comment_time = format(new Date(), 'H:mm do MMM yyyy OOOO');
-
-        const params = {
-            study_uuid: this.store.core.studyUuid,
-            user_uuid: this.store.core.userUuid,
-            history_item_index: this.store.core.studyHistoryItemIndex,
-            comment_index: index,
-            comment: comment,
-            comment_time: comment_time
-        };
-
-        this.store.core.studyHistory[
-            this.store.core.studyHistoryItemIndex
-        ].comments[index].comment = comment;
-
-        this.store.core.studyHistory[
-            this.store.core.studyHistoryItemIndex
-        ].comments[index].time = comment_time;
-
-        this.store.core.studyHistory[
-            this.store.core.studyHistoryItemIndex
-        ].comments[index]['edited'] = true;
-
-        await axios.post('history/comment/edit', params);
-
-        this.generateHistoryNodes();
     };
 
     getAllChildNodes = id => {
